requirements.txt
dlib==19.9.0
face-recognition==1.2.1
face-recognition-models==0.3.0
Flask==0.12.2
numpy==1.14.1
Pillow==5.0.0
boto3
